Fix Shopping bag bug
We intend for the minus quantity button in the bag to be disabled when
the quantity selected is 1. This works in the mobile view, but not
in the desktop view. The reason for this is that the way the page
is rendered there are two minus buttons with the same ID, it is just
that one or the other is hidden using CSS, depending on the size of
the screen.
So as far as the JavaScript code is concerned, only the first
button (the one in mobile view) has a valid ID, and the second one is
ignored, causing the desktop minus button not to be disabled when the
quantity selected is 1
One intrepid tutor (thanks Scott!) has created a fix for this,
which you can find in thefinal sourcecode for the complete project:
https://github.com/Code-Institute-Solutions/boutique_ado_v1

The details of exactly what Scott changed can be found in this commit
record :
https://github.com/Code-Institute-Solutions/boutique_ado_v1/commit/
de7ad2067ac1b5de37a4cd8b9f4ddf572a4bf6c7?diff=unified

	modified:   bag/templates/bag/quantity-form.html
	modified:   products/templates/products/
                    includes/quantity_input_script.html
	modified:   products/templates/products/product_detail.html